Move InputBase so it can be reused by other components
